requirements.txt
groq
speechrecognition
gTTS
pygame
pyaudio
python-dotenv
colorama
